requirements.txt
flask==2.0.1
werkzeug==2.0.3
paho-mqtt==1.6.1
pandas==1.3.3
numpy==1.21.2
flask-socketio==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
eventlet==0.33.0 